+++ src/test/java/org/arkecosystem/client/ConnectionManagerTest.java
@@ -9,12 +9,11 @@ import static junit.framework.TestCase.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ConnectionManagerTest {
-
     @Test
     public void connect() {
         HashMap<String, Object> map = new HashMap<>();
         map.put("host", "dummy");
-        map.put("version", 2);
+        map.put("API-Version", 1);
 
         ConnectionManager manager = new ConnectionManager();
         manager.connect(map);
@@ -25,7 +24,7 @@ public class ConnectionManagerTest {
     public void disconnect() {
         HashMap<String, Object> map = new HashMap<>();
         map.put("host", "dummy");
-        map.put("version", 2);
+        map.put("API-Version", 1);
 
         ConnectionManager manager = new ConnectionManager();
         manager.connect(map);
@@ -38,7 +37,7 @@ public class ConnectionManagerTest {
     public void connection() {
         HashMap<String, Object> map = new HashMap<>();
         map.put("host", "dummy");
-        map.put("version", 2);
+        map.put("API-Version", 1);
 
         ConnectionManager manager = new ConnectionManager();
         manager.connect(map);
@@ -68,14 +67,14 @@ public class ConnectionManagerTest {
 
         HashMap<String, Object> map = new HashMap<>();
         map.put("host", "dummy");
-        map.put("version", 2);
+        map.put("API-Version", 1);
 
         Connection<Api> connection1 = manager.connect(map);
         assertNotNull(connection1);
 
         map = new HashMap<>();
         map.put("host", "dummy");
-        map.put("version", 2);
+        map.put("API-Version", 2);
 
         Connection<Api> connection2 = manager.connect(map, "backup");
         assertNotNull(connection2);