+++ src/test/java/org/arkecosystem/client/api/TransactionsTest.java
@@ -6,6 +6,7 @@ import org.arkecosystem.client.MockHelper;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.HashMap;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -21,8 +22,8 @@ public class TransactionsTest {
 
     @Test
     void create() throws IOException {
-        Connection<Api> connection = MockHelper.connection(2);
-        LinkedTreeMap<String, Object> actual = connection.api().transactions.create(new HashMap<>());
+        Connection<Two> connection = MockHelper.connection(2);
+        LinkedTreeMap<String, Object> actual = connection.api().transactions.create(new ArrayList<>());
         assertTrue((boolean) actual.get("success"));
     }